Change QPS type to float64 (#6303)
* Change QPS type to float64


* update cmd


* lint


* fix tests


* lint


* nit


* fix tests


* fix tests


* fix tests


---------

+++ flyteplugins/go/tasks/pluginmachinery/webapi/pluginconfig_flags_test.go
@@ -105,7 +105,7 @@ func TestPluginConfig_SetFlags(t *testing.T) {
 			testValue := "1"
 
 			cmdFlags.Set("readRateLimiter.qps", testValue)
-			if vInt, err := cmdFlags.GetInt("readRateLimiter.qps"); err == nil {
+			if vInt, err := cmdFlags.GetFloat64("readRateLimiter.qps"); err == nil {
 				testDecodeJson_PluginConfig(t, fmt.Sprintf("%v", vInt), &actual.ReadRateLimiter.QPS)
 
 			} else {
@@ -133,7 +133,7 @@ func TestPluginConfig_SetFlags(t *testing.T) {
 			testValue := "1"
 
 			cmdFlags.Set("writeRateLimiter.qps", testValue)
-			if vInt, err := cmdFlags.GetInt("writeRateLimiter.qps"); err == nil {
+			if vInt, err := cmdFlags.GetFloat64("writeRateLimiter.qps"); err == nil {
 				testDecodeJson_PluginConfig(t, fmt.Sprintf("%v", vInt), &actual.WriteRateLimiter.QPS)
 
 			} else {
